+++ src/FSimBase.jl
@@ -2,17 +2,21 @@ module FSimBase
 
 
 using Reexport
-using DiffEqBase: DECallback, DEIntegrator, ODEProblem, CallbackSet, solve, isinplace
+import DiffEqBase
+using DiffEqBase: DECallback, DEIntegrator, ODEProblem, DiscreteProblem, CallbackSet, isinplace, init, solve
 using DiffEqCallbacks: SavedValues, SavingCallback
 @reexport using SimulationLogger
 using Transducers: Map
 using DataFrames: DataFrame
 using NamedTupleTools: namedtuple
+using DataFrames
 
 
 export AbstractEnv, State, Params, Dynamics, Dynamics!
 export AbstractController, Command
-export sim, apply_inputs
+export apply_inputs
+export Simulator
+export solve, reinit!, log!, step!, step_until!
 
 
 include("APIs/APIs.jl")