+++ NEWS.md
@@ -1,2 +1,6 @@
+# Changes in v0.2.0
+- The API is changed; use `simulator` for both non-interactive and interactive simulation.
+- [`DiscreteProblem`](https://diffeq.sciml.ai/stable/solvers/discrete_solve/#DiscreteProblems) can also be tested.
+
 # Changes in v0.1.1
 * The API `_sim` is changed to `sim` and exported.